added Home page and navigation page
